+++ jquery.reel.js
@@ -24,8 +24,8 @@
  *
  * jQuery Reel
  * http://jquery.vostrel.cz/reel
- * Version: 1.1.1
- * Updated: 2011-03-20
+ * Version: 1.1.1 GH-30 fix
+ * Updated: 2011-04-02
  *
  * Requires jQuery 1.4.2 or higher
  */
@@ -46,7 +46,7 @@
 jQuery.reel || (function($, window, document, undefined){
 
   $.reel= {
-    version: '1.1.1',
+    version: '1.1.1 GH-30',
 
     // Options defaults
     def: {
@@ -684,7 +684,9 @@ jQuery.reel || (function($, window, document, undefined){
   // PRIVATE
   var
     pool= $(document),
+    browser_version= +$.browser.version.split('.').slice(0,2).join('.'),
     ie= $.browser.msie,
+    knows_data_url= !(ie && browser_version < 8),
     failsafe_cursor= 'ew-resize',
     ticker,
     ticks= { before: 0, now: new Date() },
@@ -698,7 +700,7 @@ jQuery.reel || (function($, window, document, undefined){
     hi_klass= klass + '-interface',
 
     // Embedded images
-    transparent= 'data:image/gif;base64,R0lGODlhCAAIAIAAAAAAAAAAACH5BAEAAAAALAAAAAAIAAgAAAIHhI+py+1dAAA7',
+    transparent= knows_data_url ? 'data:image/gif;base64,R0lGODlhCAAIAIAAAAAAAAAAACH5BAEAAAAALAAAAAAIAAgAAAIHhI+py+1dAAA7' : cdn('blank.gif'),
     drag_cursor= 'data:image/gif;base64,R0lGODlhEAAQAJECAAAAAP///////wAAACH5BAEAAAIALAAAAAAQABAAQAI3lC8AeBDvgosQxQtne7yvLWGStVBelXBKqDJpNzLKq3xWBlU2nUs4C/O8cCvU0EfZGUwt19FYAAA7',
     drag_cursor_down= 'data:image/gif;base64,R0lGODlhEAAQAJECAAAAAP///////wAAACH5BAEAAAIALAAAAAAQABAAQAIslI95EB3MHECxNjBVdE/5b2zcRV1QBabqhwltq41St4hj5konmVioZ6OtEgUAOw==',
 
@@ -732,6 +734,7 @@ jQuery.reel || (function($, window, document, undefined){
   // Helpers
   function tag(string){ return '<' + string + '/>' }
   function dot(string){ return '.' + string }
+  function cdn(path){ return 'http://code.vostrel.cz/' + path }
   function url(location){ return 'url(' + location + ')' }
   function lofi(number){ return +number.toFixed(4) }
   function min_max(minimum, maximum, number){ return max(minimum, min(maximum, number)) }