+++ CHANGELOG.markdown
@@ -5,6 +5,7 @@ Development
 -----------
 
 * Fixed two issues preventing Reel from animating itself in IE (Github issues GH-30).
+* Fixed broken-image-overlay bug when using sprite on IE 7 and lower.
 
 
 Version 1.1.1